requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
pillow>=10.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
torch>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

